fix: exclude the use state and make the logic inside the fetch component for rastreamento
