+++ modernize/src/app/(DashboardLayout)/apps/orcamento/backoffice/components/useGetPedidoOrcamento.ts
@@ -1,4 +1,31 @@
+import getBrazilTime from "@/utils/brazilTime";
+import { isAfter, isEqual } from "date-fns";
 import { useState, useEffect, useMemo } from "react";
+
+interface Orcamento {
+  id: number;
+  user_id: number;
+  cliente_octa_number: string;
+  nome_cliente: string | null;
+  lista_produtos: string | null;
+  produtos_brinde: string | null;
+  texto_orcamento: string | null;
+  endereco_cep: string;
+  endereco: string;
+  opcao_entrega: string;
+  prazo_opcao_entrega: number;
+  preco_opcao_entrega: number | null;
+  status: string;
+  created_at: string;
+  updated_at: string;
+  brinde: number;
+  tipo_desconto: string;
+  valor_desconto: number;
+  data_antecipa: string;
+  taxa_antecipa: string;
+  total_orcamento: number;
+  prazo_producao: number;
+}
 
 interface Pedido {
   id: number;
@@ -25,9 +52,10 @@ interface Pedido {
 
 const useFetchPedidoOrcamento = (id: number | string) => {
   const [pedido, setPedido] = useState<Pedido | null>(null);
+  const [orcamento, setOrcamento] = useState<Orcamento | null>(null);
   const [isLoadingPedido, setIsLoading] = useState(true);
   const [errorPedido, setError] = useState<Error | null>(null);
-  const [activeStep, setActiveStep] = useState<number | undefined>(1);
+  const [activeStep, setActiveStep] = useState<number | null>(null);
 
 
   useEffect(() => {
@@ -42,7 +70,7 @@ const useFetchPedidoOrcamento = (id: number | string) => {
         if (!accessToken) throw new Error("Usuário não autenticado.");
 
         const res = await fetch(
-          `${process.env.NEXT_PUBLIC_API}/api/pedidos/get-pedido-orcamento/${id}`,
+          `${process.env.NEXT_PUBLIC_API}/api/pedidos/get-pedido-&-orcamento/${id}`,
           {
             method: "GET",
             headers: {
@@ -55,17 +83,33 @@ const useFetchPedidoOrcamento = (id: number | string) => {
         if (!res.ok) throw new Error("Erro ao buscar orçamentos.");
 
         const json = await res.json();
-        console.log(json);
-        setPedido(json);
-        if(json.pedido_status_id === 14){
+        console.log(json.orcamento);
+        setPedido(json.pedido);
+        setOrcamento(json.orcamento);
+
+        const createdAtOrcamento = orcamento?.created_at ? new Date(orcamento.created_at) : null;
+        const createdPedido = pedido?.created_at ? new Date(pedido.created_at) : null;
+        const hoje = getBrazilTime();
+
+        if (createdAtOrcamento && (isAfter(hoje, createdAtOrcamento) || isEqual(hoje, createdAtOrcamento))) {
+          setActiveStep(2);
+        }
+        
+        if (createdPedido && (isAfter(hoje, createdPedido) || isEqual(hoje, createdPedido))) {
           setActiveStep(3);
         }
-        if(json.pedido_status_id === 15){
+
+        if (json.pedido.pedido_status_id === 24 || json.pedido.pedido_status_id === 23) {
+          setActiveStep(3);
+          if (json.pedido.pedido_status_id === 25) {
+        }
           setActiveStep(4);
         }
+        
       } catch (err) {
         setError(err as Error);
       } finally {
+        console.log(activeStep)
         setIsLoading(false);
       }
     };
@@ -74,7 +118,7 @@ const useFetchPedidoOrcamento = (id: number | string) => {
   }, [id]);
 
   // Retorna um objeto memorizado para otimizar performance
-  return useMemo(() => ({ pedido, isLoadingPedido, errorPedido, activeStep }), [pedido, isLoadingPedido, errorPedido, activeStep]);
+  return useMemo(() => ({ orcamento, pedido, isLoadingPedido, errorPedido, activeStep }), [orcamento, pedido, isLoadingPedido, errorPedido, activeStep]);
 };
 
 export default useFetchPedidoOrcamento;
